Upgrade ffmpeg_kit_flutter_min_gpl to 6.0.3

+++ lib/extensions/statistics_extension.dart
@@ -3,7 +3,7 @@ import 'package:ffmpeg_helper/ffmpeg_helper.dart';
 extension OnStatistics on Statistics {
   /// get time by duration
   Duration getTimeDuration() {
-    return Duration(microseconds: getTime());
+    return Duration(microseconds: getTime().toInt());
   }
 
   /// get time by milliseconds
