+++ lib/helpers/ffmpeg_helper_class.dart
@@ -199,7 +199,7 @@ class FFMpegHelper {
             double.tryParse(temp['fps']) ?? 0.0,
             double.tryParse(temp['stream_0_0_q']) ?? 0.0,
             int.tryParse(temp['total_size']) ?? 0,
-            int.tryParse(temp['out_time_us']) ?? 0,
+            double.tryParse(temp['out_time_us']) ?? 0.0,
             // 2189.6kbits/s => 2189.6
             double.tryParse(
                     temp['bitrate']?.replaceAll(RegExp('[a-z/]'), '')) ??
